pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apify-client>=1.12.0",
    "asgiref>=3.9.1",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "openai>=1.93.2",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]
